refactor(server): extract Cli and introspection helpers from main.rs
- New crates/fakecloud-server/src/cli.rs holds the Cli struct and a
  small endpoint_url() helper that derives the public URL from the
  configured bind address (rewriting wildcard hosts to localhost).
- New crates/fakecloud-server/src/introspection.rs hosts the four
  introspection response builders that translate raw service state
  into the public RdsInstance / ElastiCache* SDK types, plus the
  existing test that pins the password-redaction contract.
- main.rs now wires these in via mod cli; mod introspection; and
  shrinks by 165 lines, leaving only the bind/setup orchestration.
